+++ src/instructions/andi.java
@@ -17,7 +17,7 @@ public class andi extends IFormat {
         int v1 = Register.getRegister().readRegister(getRt());
         int v2 = getConstant();
         int result = ALU.executeInstruction("and", v1, v2);
-        Register.getRegister().writeRegister(getRs(), (byte) result);
+        Register.getRegister().writeRegister(getRs(), result);
     }
 }
